moved some code to gqlgen package
